refactor: make decomposition
- Add global forwards.
- Add experimental factory class.
- Split asynchronous sink into several files.
- Implement type method for asynchronous sink factory.
