+++ include/blackhole/forward.hpp
@@ -5,8 +5,22 @@ inline namespace v1 {
 
 class severity_t;
 
-template<typename>
+class formatter_t;
+class handler_t;
+class registry_t;
+class sink_t;
+
+template<typename T>
 struct factory;
+
+class factory_t;
+
+namespace experimental {
+
+template<typename T>
+class factory;
+
+}  // namespace experimental
 
 namespace config {
